# joysong-server/openapi.yaml
openapi: 3.0.3
info:
  title: 娇颜颂 用户端 API
  description: 娇颜颂医美平台 Android App 用户端接口文档（不含管理后台）
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Token 认证，Header: Authorization: Bearer <token>'
  schemas:
    BaseResponse:
      type: object
      description: 统一响应包装
      properties:
        code:
          type: integer
          example: 200
          description: 状态码，200=成功
        message:
          type: string
          example: success
        data:
          description: 响应数据，类型随接口而定
    LoginRequest:
      type: object
      required:
      - phone
      - password
      properties:
        phone:
          type: string
          example: '13800138000'
          description: 手机号
        password:
          type: string
          example: '123456'
          description: 密码
    RegisterRequest:
      type: object
      required:
      - phone
      - password
      properties:
        phone:
          type: string
          example: '13800138000'
        code:
          type: string
          example: '1234'
          description: 验证码
        password:
          type: string
          example: '123456'
    SendCodeRequest:
      type: object
      required:
      - phone
      properties:
        phone:
          type: string
          example: '13800138000'
    LoginWithCodeRequest:
      type: object
      required:
      - phone
      - code
      properties:
        phone:
          type: string
          example: '13800138000'
        code:
          type: string
          example: '123456'
          description: 验证码
    GoogleLoginRequest:
      type: object
      required:
      - idToken
      properties:
        idToken:
          type: string
          description: Google ID Token（由客户端 Credential Manager 获取）
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6Ij...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzM4NCJ9...
          description: JWT Token
        user:
          $ref: '#/components/schemas/UserDto'
    UserDto:
      type: object
      description: 用户信息
      properties:
        id:
          type: string
          format: uuid
          example: c233b71f-5359-414e-8dd8-8deb29a60f92
        phone:
          type: string
          nullable: true
          example: '13800138000'
        email:
          type: string
          nullable: true
          example: user@gmail.com
        nickname:
          type: string
          example: 小美
        avatar:
          type: string
          example: https://joysong--images.oss-cn-hangzhou.aliyuncs.com/avatars/xxx.jpg
        gender:
          type: string
          example: 女
        city:
          type: string
          nullable: true
          example: 上海
        bio:
          type: string
          nullable: true
          example: 喜欢尝试皮肤管理项目。
        birthday:
          type: string
          nullable: true
          example: '1995-06-15'
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
          nullable: true
          example: 小美
        avatar:
          type: string
          nullable: true
          example: https://...
        gender:
          type: string
          nullable: true
          example: 女
        city:
          type: string
          nullable: true
          example: 上海
        bio:
          type: string
          nullable: true
          example: 喜欢尝试皮肤管理项目。
        birthday:
          type: string
          nullable: true
          example: '1995-06-15'
...
    AccountDeletionBlocker:
      type: object
      required: [type, count, action]
      description: 不包含订单号、申请号或其他敏感业务 ID 的阻断摘要
      properties:
        type: { type: string, example: IDENTITY_APPLICATION }
        count: { type: integer, format: int64, minimum: 1, example: 1 }
        action: { type: string, example: VIEW_IDENTITY_APPLICATION }
    AccountDeletionPreflightResponse:
      type: object
      required: [requestId, eligible, policyVersion, blockers]
      properties:
        requestId: { type: string, format: uuid }
        eligible: { type: boolean }
        stepUpMethod: { type: string, nullable: true, enum: [SMS, GOOGLE] }
        maskedCredential: { type: string, nullable: true, example: '+86****00' }
        policyVersion: { type: string, example: '2026-08-28' }
        blockers: { type: array, items: { $ref: '#/components/schemas/AccountDeletionBlocker' } }
    AccountDeletionStepUpRequest:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string, format: uuid }
        code: { type: string, pattern: '^[0-9]{6}$', nullable: true }
        googleIdToken: { type: string, nullable: true, writeOnly: true }
    AccountDeletionSmsCodeResponse:
      type: object
      required: [requestId, expiresInSeconds, resendAfterSeconds]
      properties:
        requestId: { type: string, format: uuid }
        expiresInSeconds: { type: integer, format: int64, example: 300 }
        resendAfterSeconds: { type: integer, format: int64, example: 60 }
    AccountDeletionStepUpResponse:
      type: object
      required: [requestId, deletionAuthorization, expiresInSeconds]
      properties:
        requestId: { type: string, format: uuid }
        deletionAuthorization: { type: string, readOnly: true, description: 一次性授权，服务端仅保存 HMAC }
        expiresInSeconds: { type: integer, format: int64, example: 600 }
    ConfirmAccountDeletionRequest:
      type: object
      required: [requestId, policyVersion, confirmation]
      properties:
        requestId: { type: string, format: uuid }
        policyVersion: { type: string, example: '2026-08-28' }
        confirmation: { type: string, enum: [DELETE] }
    AccountDeletionConfirmResponse:
      type: object
      required: [requestId, outcome, completedAt, blockers]
      properties:
        requestId: { type: string, format: uuid }
        outcome: { type: string, enum: [ERASED, BLOCKED] }
        completedAt: { type: string, format: date-time }
        blockers: { type: array, items: { $ref: '#/components/schemas/AccountDeletionBlocker' } }
paths:
  /api/user/account-deletion/preflight:
    post:
      tags: [用户模块]
      summary: 创建账号注销预检请求
      operationId: accountDeletionPreflight
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: 预检完成；业务阻断仍返回 200 且 eligible=false
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties: { data: { $ref: '#/components/schemas/AccountDeletionPreflightResponse' } }
        '401': { description: 未认证，ACCOUNT_DELETION_REQUEST_FAILED }
        '503': { description: 功能关闭或交易阻断服务不可用 }
  /api/user/account-deletion/send-sms-code:
    post:
      tags: [用户模块]
      summary: 发送账号注销短信验证码
      operationId: sendAccountDeletionSmsCode
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId]
              properties: { requestId: { type: string, format: uuid } }
      responses:
        '200':
          description: 已发送；验证码 300 秒有效，60 秒后可重发
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties: { data: { $ref: '#/components/schemas/AccountDeletionSmsCodeResponse' } }
        '400': { description: 请求不适用于 SMS 验证 }
        '401': { description: 未认证 }
        '429': { description: ACCOUNT_DELETION_SMS_RATE_LIMITED }
        '503': { description: ACCOUNT_DELETION_SMS_DELIVERY_UNAVAILABLE 或功能关闭 }
  /api/user/account-deletion/step-up:
    post:
      tags: [用户模块]
      summary: 完成账号注销二次验证
      operationId: accountDeletionStepUp
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountDeletionStepUpRequest' }
      responses:
        '200':
          description: 返回 600 秒有效的一次性注销授权
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties: { data: { $ref: '#/components/schemas/AccountDeletionStepUpResponse' } }
        '400': { description: ACCOUNT_DELETION_VERIFICATION_FAILED }
        '401': { description: 未认证 }
        '410': { description: ACCOUNT_DELETION_AUTHORIZATION_EXPIRED }
        '503': { description: 功能关闭或 Google 验证服务不可用 }
  /api/user/account-deletion/confirm:
    post:
      tags: [用户模块]
      summary: 幂等确认不可逆账号注销
      operationId: confirmAccountDeletion
      description: 首次调用必须携带 Bearer；注销后允许使用完全相同的 requestId、幂等键和授权查询原终态。
      security: [{ BearerAuth: [] }, {}]
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema: { type: string, minLength: 16, maxLength: 256 }
      - name: X-Account-Deletion-Authorization
        in: header
        required: true
        schema: { type: string, minLength: 32, maxLength: 512, writeOnly: true }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmAccountDeletionRequest' }
      responses:
        '200':
          description: 注销成功或完全匹配的 ERASED 终态重取
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties: { data: { $ref: '#/components/schemas/AccountDeletionConfirmResponse' } }
        '400': { description: ACCOUNT_DELETION_REQUEST_FAILED }
        '401': { description: 首次确认缺少有效 Bearer }
        '409': { description: ACCOUNT_DELETION_BLOCKED 或 ACCOUNT_DELETION_IDEMPOTENCY_CONFLICT }
        '410': { description: ACCOUNT_DELETION_AUTHORIZATION_EXPIRED }
        '503': { description: 功能关闭或阻断服务不可用 }
...
  /api/auth/login:
    post:
      tags:
      - 认证模块
      summary: 手机号密码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/LoginResponse'
  /api/auth/register:
    post:
      tags:
      - 认证模块
      summary: 手机号注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/LoginResponse'
  /api/auth/send-code:
    post:
      tags:
      - 认证模块
      summary: 发送验证码（Mock）
      operationId: sendCode
      description: 开发环境直接返回验证码，生产环境通过短信服务发送
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        message:
                          type: string
                          example: 验证码已发送
                        code:
                          type: string
                          example: '123456'
  /api/auth/login-with-code:
    post:
      tags:
      - 认证模块
      summary: 验证码登录
      operationId: loginWithCode
      description: 用户不存在时自动注册（昵称默认为手机号）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginWithCodeRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/LoginResponse'
  /api/auth/login-with-google:
    post:
      tags:
      - 认证模块
      summary: Google 登录
      operationId: loginWithGoogle
      description: 使用 Google ID Token 登录。服务端通过 Google tokeninfo 端点验证 Token 有效性，并校验 audience 字段。用户不存在时自动注册。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/LoginResponse'
  /api/user/profile:
    get:
      tags:
      - 用户模块
      summary: 获取当前用户信息
      operationId: getProfile
      security:
      - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserDto'
    put:
      tags:
      - 用户模块
      summary: 更新当前用户信息
      operationId: updateProfile
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserDto'
  /api/user/avatar:
    put:
      tags:
      - 用户模块
      summary: 更新头像
      operationId: updateAvatar
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - avatar
              properties:
                avatar:
                  type: string
                  example: https://joysong--images.oss-cn-hangzhou.aliyuncs.com/avatars/xxx.jpg
                  description: 头像URL
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/UserDto'
...
  /api/user/password:
    put:
      tags:
      - 用户模块
      summary: 修改密码
      operationId: changePassword
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - oldPassword
              - newPassword
              properties:
                oldPassword:
                  type: string
                  example: oldPass123
                  description: 原密码
                newPassword:
                  type: string
                  example: newPass456
                  description: 新密码（至少6位）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/BaseResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        message:
                          type: string
                          example: 密码修改成功
...